Support standard ortho4x12 QMK layout
QMK allows layouts to be shared across boards with the same physical
layouts (e.g. planck and houndstooth), which is really useful if you own
more than one board and want to have the same key mappings on each.

Supporting is straightforward, just the declaration in rules.mk and a
layout declaration which matches the convention. It doesn't affect
builds which don't want to use it, and which just create a new keymap.c.

https://beta.docs.qmk.fm/features/feature_layouts